Update various things
In particular, update nix-darwin to a version that has release channel
branches. This seems to have broken karabiner elements, thought it never
actually did what I wanted in the first place.

There's a temporary hack to deal with the fact that my neovim setup
doesn't work with the latest version of the neovim flake, and that the
neovim flake version I was using doesn't build with the latest nixpkgs.